:truck: Update and rename 'getImage' to 'fallbackImageUrl'
