+++ aksel.nav.no/website/components/utils/index.ts
@@ -1,4 +1,3 @@
 export * from "./generate-sidebar";
 export * from "./generate-tableofcontents";
 export * from "./get-authors";
-export * from "./get-image";
